move configview to component
